More moves and rebase fixes
