+++ distarray/local/construct.py
@@ -8,7 +8,7 @@ from __future__ import division
 
 from distarray.local.mpiutils import MPI
 from distarray.local.error import NullCommError, InvalidBaseCommError
-from distarray import mpiutils
+from distarray.local import mpiutils
 
 
 # ---------------------------------------------------------------------------
