+++ distarray/tests/test_testing.py
